+++ src/main/locales/da/locale.h
@@ -93,8 +93,8 @@
 #define TR_CLI_DSHOT_READ           "Dshot laest: %u"
 #define TR_CLI_DSHOT_INVALID_PKT    "Dshot ugyldig pakker: %u"
 #define TR_CLI_DSHOT_DIR_CHANGE     "Dshot retning  Skift  cyklus: %u, mikro:  %u"
-#define TR_CLI_DSHOT_HEAD1          "Motor    Type   eRPM    RPM     Hz    Invalid   TEMP    VCC   CURR  ST/EV   DBG1   DBG2   DBG3"
-#define TR_CLI_DSHOT_LINE1          "=====  ====== ====== ====== ====== ========== ====== ====== ====== ====== ====== ====== ======"
+#define TR_CLI_DSHOT_HEAD1          "Motor    Type   eRPM    RPM     Hz Invalid   TEMP    VCC   CURR  ST/EV   DBG1   DBG2   DBG3"
+#define TR_CLI_DSHOT_LINE1          "=====  ====== ====== ====== ====== ======= ====== ====== ====== ====== ====== ====== ======"
 #define TR_CLI_DSHOT_HEAD2          "Motor    Type   eRPM    RPM     Hz   TEMP    VCC   CURR  ST/EV   DBG1   DBG2   DBG3"
 #define TR_CLI_DSHOT_LINE2          "=====  ====== ====== ====== ====== ====== ====== ====== ====== ====== ====== ======"
 #define TR_CLI_DSHOT_NO_TELEM       "Dshot telemetri ikke aktiv"
